+++ examples/uuid/src/main.rs
@@ -1,50 +1,67 @@
-#[macro_use]
-extern crate rocket;
-#[macro_use]
-extern crate rocket_okapi;
-
-use rocket::serde::{json::Json, uuid::Uuid};
-use rocket_okapi::swagger_ui::*;
+use rocket::{get, post, serde::json::Json, serde::uuid::Uuid};
+use rocket_okapi::settings::UrlObject;
+use rocket_okapi::{openapi, openapi_get_routes, rapidoc::*, swagger_ui::*};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, JsonSchema)]
 #[serde(rename_all = "camelCase")]
 struct User {
+    #[schemars(example = "example_uuid")]
     user_id: uuid::Uuid,
+    username: String,
+    #[schemars(example = "example_email")]
+    email: Option<String>,
+}
+
+fn example_email() -> &'static str {
+    "test@example.com"
+}
+fn example_uuid() -> &'static str {
+    "fdb12d51-0e3f-4ff8-821e-fbc255d8e413"
 }
 
 /// # Get all users
 ///
 /// Returns all users in the system.
-#[openapi]
+#[openapi(tag = "Users")]
 #[get("/user")]
 fn get_all_users() -> Json<Vec<User>> {
     Json(vec![User {
         user_id: uuid::Uuid::new_v4(),
+        username: "bob".to_owned(),
+        email: None,
     }])
 }
 
 /// # Get user
 ///
 /// Returns a single user by ID.
-#[openapi]
+#[openapi(tag = "Users")]
 #[get("/user/<id>")]
 fn get_user(id: Uuid) -> Option<Json<User>> {
-    Some(Json(User { user_id: id }))
+    Some(Json(User {
+        user_id: id,
+        username: "bob".to_owned(),
+        email: None,
+    }))
 }
 
 /// # Get user by name
 ///
 /// Returns a single user by username.
-#[openapi]
-#[get("/user_example?<id>")]
-fn get_user_by_name(id: Uuid) -> Option<Json<User>> {
-    Some(Json(User { user_id: id }))
+#[openapi(tag = "Users")]
+#[get("/user_example?<user_id>&<name>&<email>")]
+fn get_user_by_name(user_id: Uuid, name: String, email: Option<String>) -> Option<Json<User>> {
+    Some(Json(User {
+        user_id,
+        username: name,
+        email,
+    }))
 }
 
 /// # Create user
-#[openapi]
+#[openapi(tag = "Users")]
 #[post("/user", data = "<user>")]
 fn create_user(user: Json<User>) -> Json<User> {
     user
@@ -52,10 +69,10 @@ fn create_user(user: Json<User>) -> Json<User> {
 
 #[rocket::main]
 async fn main() {
-    let result = rocket::build()
+    let launch_result = rocket::build()
         .mount(
             "/",
-            routes_with_openapi![get_all_users, get_user, get_user_by_name, create_user,],
+            openapi_get_routes![get_all_users, get_user, get_user_by_name, create_user,],
         )
         .mount(
             "/swagger-ui/",
@@ -64,7 +81,25 @@ async fn main() {
                 ..Default::default()
             }),
         )
+        .mount(
+            "/rapidoc/",
+            make_rapidoc(&RapiDocConfig {
+                general: GeneralConfig {
+                    spec_urls: vec![UrlObject::new("General", "../openapi.json")],
+                    ..Default::default()
+                },
+                hide_show: HideShowConfig {
+                    allow_spec_url_load: false,
+                    allow_spec_file_load: false,
+                    ..Default::default()
+                },
+                ..Default::default()
+            }),
+        )
         .launch()
         .await;
-    assert!(result.is_ok());
+    match launch_result {
+        Ok(()) => println!("Rocket shut down gracefully."),
+        Err(err) => println!("Rocket had an error: {}", err),
+    };
 }